requirements.txt
django>=4.1.2
django-cors-headers==3.14.0
django-debug-toolbar==3.6.0
django-extensions==3.2.1
djangorestframework==3.14.0
psycopg2==2.9.3
python-dotenv==0.21.0
pytils==0.4.1